weather_air_quality_etl_pipeline/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
greenlet==3.1.0
idna==3.10
numpy==2.0.2
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3

# Flask API
Flask==3.0.3
flask-cors==4.0.0
APScheduler==3.10.4
